WPT test

+++ R4_STM_CODE/Core/Inc/hx711.hh
@@ -18,7 +18,7 @@ public:
 	int32_t getBitsToGramRatio() const;
 	void setBitsToGramRatio(int32_t newCalibrationFactor);
 	void addToOffsetInGrams(int32_t offsetDif);
-	void initialCalibration(uint8_t testWeightInGrams = 1000);
+	void initialCalibration(uint32_t testWeightInGrams = 1000);
 	int32_t ReadValue();
 	int32_t AverageValue(uint16_t times = 10);
 };
